refactor : 이미지 관련 코드 메인 액티비티로 배치
