+++ TripFriend/app/src/main/java/com/test/tripfriend/ui/main/MainActivity.kt
@@ -1,13 +1,24 @@
 package com.test.tripfriend.ui.main
 
+import android.app.Activity
+import android.content.Intent
+import android.graphics.Bitmap
+import android.graphics.BitmapFactory
+import android.graphics.Matrix
+import android.media.ExifInterface
+import android.net.Uri
 import android.os.Bundle
 import android.os.SystemClock
 import android.view.View
 import android.view.inputmethod.InputMethodManager
+import android.widget.ImageView
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentTransaction
+import com.bumptech.glide.Glide
 import com.test.tripfriend.R
 import com.test.tripfriend.databinding.ActivityMainBinding
 import com.test.tripfriend.ui.accompany.AccompanyRegisterFragment1
@@ -190,5 +201,51 @@ class MainActivity : AppCompatActivity() {
 
             inputMethodManager.hideSoftInputFromWindow(currentFocus?.windowToken, InputMethodManager.HIDE_NOT_ALWAYS)
         }
+    }
+
+    //uri를 이미지뷰에 셋팅하는 함수
+    fun setImage(image: Uri, imageView: ImageView){
+        val inputStream = contentResolver.openInputStream(image)
+        val bitmap = BitmapFactory.decodeStream(inputStream)
+
+        //회전 각도값을 가져옴
+        val degree = getDegree(image)
+
+        //회전 이미지를 생성한다
+        val matrix = Matrix()
+        matrix.postRotate(degree.toFloat())
+        val rotateBitmap = Bitmap.createBitmap(bitmap!!, 0, 0, bitmap.width, bitmap.height, matrix, false)
+
+        //글라이드 라이브러리로 view에 이미지 출력
+        Glide.with(this).load(rotateBitmap)
+            .into(imageView)
+    }
+
+    // 이미지 파일에 기록되어 있는 회전 정보를 가져온다.
+    fun getDegree(uri: Uri) : Int{
+        var exifInterface: ExifInterface? = null
+
+        // 사진 파일로 부터 tag 정보를 관리하는 객체를 추출한다.
+        try {
+            val inputStream = contentResolver.openInputStream(uri)
+            if (inputStream != null) {
+                exifInterface = ExifInterface(inputStream)
+            }
+        } catch (e: Exception) {
+            e.printStackTrace()
+        }
+
+        var degree = 0
+        if(exifInterface != null){
+            // 각도 값을 가지고온다.
+            val orientation = exifInterface.getAttributeInt(ExifInterface.TAG_ORIENTATION, -1)
+
+            when(orientation){
+                ExifInterface.ORIENTATION_ROTATE_90 -> degree = 90
+                ExifInterface.ORIENTATION_ROTATE_180 -> degree = 180
+                ExifInterface.ORIENTATION_ROTATE_270 -> degree = 270
+            }
+        }
+        return degree
     }
 }
